[Bronze V] Title: 꼬마 정민, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ 백준/Bronze/11382. 꼬마 정민/꼬마 정민.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-long int main() {
+int main() {
     long int a, b, c ;
     
     scanf("%ld %ld %ld", &a, &b, &c);
